requirements.txt
numpy==1.18.2
pandas==1.0.3
geopandas==0.7.0
faiss-cpu==1.6.3
tqdm==4.45.0
flask==1.1.2
shapely==1.7.0
pathos==0.2.5
requests==2.23.0
sklearn==0.0

torch==1.4.0
torchvision==0.5.0
opencv-python==4.2.0.32
moviepy==1.0.1
matplotlib==3.0.3

wandb==0.8.29

jupyter==1.0.0
jupyterlab==1.2.1

colorlog==4.0.2
